Implementando el patron CircuitBreaker y entornos de desarrolo

+++ src/main/java/com/microservicios/springboot/app/item/models/Item.java
@@ -14,6 +14,6 @@ public class Item {
     private Integer Cantidad;
 
     public Double getTotal(){
-        return producto.getPrecio() * Cantidad.doubleValue();
+        return producto.getPrecio() * Cantidad;
     }
 }
